Updating OutsideClickHandler to accept "flex" as a display prop
There are tons of different ways to handle the ternary for the style property on the main `div`, so feel free to chime in with something else if you think it would be more efficient.

Performing requested updates to patch-1

Updating the README.md to match the addition of flex

README formatting issue

Adding a missing quite to README

Reverting formatting changes

Reverting formatting changes brought on by text editor using prettier.

Bold to italics

Reverting bold to italics in the README